Fix change detection and add some quality of life
